api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
shapely==2.0.2
Pillow==10.1.0
reportlab==4.0.7
requests==2.31.0
gunicorn==21.2.0
pvlib==0.10.2
pandas==2.1.3